backend_taskforge_ai/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4
alembic
passlib[bcrypt]
python-jose>=3.3.0
python-multipart
aiofiles
pytest
httpx
